add new folder for divide the functions and router.Refactor the code
